[#47] [telepath-ios] Converts SecureChannel to a protocol
