--- workspaces/telepath-ios/Telepath/ChannelKey.swift
@@ -1,3 +1,0 @@
-import Sodium
-
-public typealias ChannelKey = SecretBox.Key
